+++ ckb/src/main/java/org/nervos/ckb/service/RpcResponse.java
@@ -1,0 +1,14 @@
+package org.nervos.ckb.service;
+
+/** Copyright © 2020 Nervos Foundation. All rights reserved. */
+public class RpcResponse<T> {
+  public long id;
+  public String jsonrpc;
+  public T result;
+  public Error error;
+
+  static class Error {
+    public int code;
+    public String message;
+  }
+}