+++ example/src/main/java/org/nervos/ckb/SUDTExample.java
@@ -83,7 +83,6 @@ public class SUDTExample {
     txBuilder.addCellDep(new CellDep(new OutPoint(SUDT_OUT_POINT_TX_HASH, "0x0"), CellDep.CODE));
 
     // You can get fee rate by rpc or set a simple number
-    // BigInteger feeRate = Numeric.toBigInt(api.estimateFeeRate("5").feeRate);
     BigInteger feeRate = BigInteger.valueOf(1024);
 
     // initial_length = 2 * secp256k1_signature_byte.length
@@ -147,7 +146,6 @@ public class SUDTExample {
     txBuilder.addCellDep(new CellDep(new OutPoint(SUDT_OUT_POINT_TX_HASH, "0x0"), CellDep.CODE));
 
     // You can get fee rate by rpc or set a simple number
-    // BigInteger feeRate = Numeric.toBigInt(api.estimateFeeRate("5").feeRate);
     BigInteger feeRate = BigInteger.valueOf(1024);
 
     // initial_length = 2 * secp256k1_signature_byte.length
@@ -210,7 +208,6 @@ public class SUDTExample {
     txBuilder.addCellDep(new CellDep(new OutPoint(SUDT_OUT_POINT_TX_HASH, "0x0"), CellDep.CODE));
 
     // You can get fee rate by rpc or set a simple number
-    // BigInteger feeRate = Numeric.toBigInt(api.estimateFeeRate("5").feeRate);
     BigInteger feeRate = BigInteger.valueOf(1024);
 
     // initial_length = 2 * secp256k1_signature_byte.length